fixed bug with end date
